+++ app/delete_message.php
@@ -1,6 +1,11 @@
 <?php
+session_start(); 
+if (!isset($_SESSION['admin'])) {  // if admin is not logged in
+    header('location:  ../admin_login.php');
+  exit;
+}
     // Include the database configuration file
-    include 'database.php';
+    require('../env/database.php');
     $id = $_GET['Id']; // Get the id from the URL         
     $sql = mysqli_query($conn,"DELETE FROM contact WHERE id='$id'"); // Delete the message from contact table
     // Redirect to the message page
